refactor: move mainthread impl into mainthread-api packages
we're going to implement thread mode: `all-on-ui`
