+++ .changeset/ready-months-type.md
@@ -1,0 +1,7 @@
+---
+"@lynx-js/web-mainthread-apis": patch
+"@lynx-js/web-worker-runtime": patch
+"@lynx-js/web-constants": patch
+---
+
+refactor: move mainthread impl into mainthread-api packages
