+++ packages/web-platform/web-constants/src/index.ts
@@ -6,4 +6,5 @@ export * from './constants.js';
 export * from './eventName.js';
 export * from './endpoints.js';
 export * from './types/index.js';
+export * from './utils/index.js';
 export type * from '@lynx-js/web-worker-rpc';
