+++ packages/web-platform/web-constants/src/utils/index.ts
@@ -1,0 +1,1 @@
+export { LynxCrossThreadContext } from './LynxCrossThreadContext.js';
