+++ packages/web-platform/web-mainthread-apis/src/crossThreadHandlers/registerCallLepusMethodHandler.ts
@@ -1,9 +1,8 @@
 // Copyright 2023 The Lynx Authors. All rights reserved.
 // Licensed under the Apache License Version 2.0 that can be found in the
 // LICENSE file in the root directory of this source tree.
-import { callLepusMethodEndpoint } from '@lynx-js/web-constants';
-import type { MainThreadRuntime } from '@lynx-js/web-mainthread-apis';
-import type { Rpc } from '@lynx-js/web-worker-rpc';
+import { callLepusMethodEndpoint, type Rpc } from '@lynx-js/web-constants';
+import type { MainThreadRuntime } from '../MainThreadRuntime.js';
 
 export function registerCallLepusMethodHandler(
   rpc: Rpc,
