+++ packages/web-platform/web-mainthread-apis/src/crossThreadHandlers/registerGetCustomSectionHandler.ts
@@ -1,10 +1,10 @@
 // Copyright 2023 The Lynx Authors. All rights reserved.
 // Licensed under the Apache License Version 2.0 that can be found in the
 // LICENSE file in the root directory of this source tree.
-import type { Rpc } from '@lynx-js/web-worker-rpc';
 import {
   getCustomSectionsEndpoint,
   type LynxTemplate,
+  type Rpc,
 } from '@lynx-js/web-constants';
 
 export function registerGetCustomSectionHandler(
