+++ packages/web-platform/web-mainthread-apis/src/index.ts
@@ -2,4 +2,5 @@
 // Licensed under the Apache License Version 2.0 that can be found in the
 // LICENSE file in the root directory of this source tree.
 
+export { loadMainThread } from './loadMainThread.js';
 export * from './MainThreadRuntime.js';
